+++ src/extensions/mod.rs
@@ -16,7 +16,7 @@ pub mod merge_subscription;
 pub mod rate_limit;
 pub mod server;
 pub mod telemetry;
-pub mod validate;
+pub mod validator;
 
 #[async_trait]
 pub trait Extension: Sized {
@@ -139,5 +139,5 @@ define_all_extensions! {
     server: server::SubwayServerBuilder,
     event_bus: event_bus::EventBus,
     rate_limit: rate_limit::RateLimitBuilder,
-    validate: validate::Validate,
+    validator: validator::Validator,
 }
